# Check bucket series time range when decoding chunks from index