@bwplotka 2020-10-28
Running benchmarks is tricky but there are ways to make it reliable and more trustworthy.

* If you run Goland ID you have benchtime only 1s, so this one thing to improve is to increase this time. You can to that by edit config and setting extra program `-test.benchtime=10s`
* The other must have flag is to either set `benchmem` or specify in the code `b.ReportAllocs()`
* Useful thing is to limit to 1 or 2 CPUs, because that's the usual value we give for Store GW.