@bwplotka 2020-10-28
Anyway good work I think it makes sense, wonder if we can clean decodeWithReq a bit more :hugs: 


Some reference for benchmarks: https://gist.github.com/bwplotka/3b853c31ed11e77c975b9df45d105d74