@yeya24 2020-10-28
I finished the benchmarks and the new results between this pr and current master are:

1. SkipChunks 

```
name                                                                  old time/op    new time/op    delta
BucketSeriesSkipChunks/1000000SeriesWith1Samples/1of1000000              423ms ± 0%     442ms ± 0%   +4.32%
BucketSeriesSkipChunks/1000000SeriesWith1Samples/10of1000000             496ms ± 0%     458ms ± 0%   -7.56%
BucketSeriesSkipChunks/1000000SeriesWith1Samples/1000000of1000000        3.39s ± 0%     3.31s ± 0%   -2.29%
BucketSeriesSkipChunks/100000SeriesWith100Samples/1of10000000           38.6ms ± 0%    36.6ms ± 0%   -5.18%
BucketSeriesSkipChunks/100000SeriesWith100Samples/100of10000000         36.1ms ± 0%    42.3ms ± 0%  +17.31%
BucketSeriesSkipChunks/100000SeriesWith100Samples/10000000of10000000     276ms ± 0%     278ms ± 0%   +0.76%
BucketSeriesSkipChunks/1SeriesWith10000000Samples/1of10000000           4.53ms ± 0%    0.13ms ± 0%  -97.06%
BucketSeriesSkipChunks/1SeriesWith10000000Samples/100of10000000         4.13ms ± 0%    0.13ms ± 0%  -96.79%
BucketSeriesSkipChunks/1SeriesWith10000000Samples/10000000of10000000    16.5ms ± 0%     0.4ms ± 0%  -97.48%

name                                                                  old alloc/op   new alloc/op   delta
BucketSeriesSkipChunks/1000000SeriesWith1Samples/1of1000000             98.2MB ± 0%    98.2MB ± 0%   -0.00%
BucketSeriesSkipChunks/1000000SeriesWith1Samples/10of1000000            98.2MB ± 0%    98.2MB ± 0%   -0.00%
BucketSeriesSkipChunks/1000000SeriesWith1Samples/1000000of1000000       1.00GB ± 0%    1.00GB ± 0%   -0.00%
BucketSeriesSkipChunks/100000SeriesWith100Samples/1of10000000           8.43MB ± 0%    8.43MB ± 0%   +0.00%
BucketSeriesSkipChunks/100000SeriesWith100Samples/100of10000000         8.43MB ± 0%    8.43MB ± 0%   -0.00%
BucketSeriesSkipChunks/100000SeriesWith100Samples/10000000of10000000    95.9MB ± 0%    95.9MB ± 0%   -0.00%
BucketSeriesSkipChunks/1SeriesWith10000000Samples/1of10000000           4.34MB ± 0%    0.17MB ± 0%  -96.00%
BucketSeriesSkipChunks/1SeriesWith10000000Samples/100of10000000         4.34MB ± 0%    0.17MB ± 0%  -96.00%
BucketSeriesSkipChunks/1SeriesWith10000000Samples/10000000of10000000    17.3MB ± 0%     0.7MB ± 0%  -96.06%

name                                                                  old allocs/op  new allocs/op  delta
BucketSeriesSkipChunks/1000000SeriesWith1Samples/1of1000000               760k ± 0%      760k ± 0%   -0.00%
BucketSeriesSkipChunks/1000000SeriesWith1Samples/10of1000000              760k ± 0%      760k ± 0%   -0.00%
BucketSeriesSkipChunks/1000000SeriesWith1Samples/1000000of1000000        8.04M ± 0%     8.04M ± 0%   -0.00%
BucketSeriesSkipChunks/100000SeriesWith100Samples/1of10000000            76.1k ± 0%     76.1k ± 0%    0.00%
BucketSeriesSkipChunks/100000SeriesWith100Samples/100of10000000          76.1k ± 0%     76.1k ± 0%   -0.00%
BucketSeriesSkipChunks/100000SeriesWith100Samples/10000000of10000000      804k ± 0%      804k ± 0%   -0.00%
BucketSeriesSkipChunks/1SeriesWith10000000Samples/1of10000000              195 ± 0%       172 ± 0%  -11.79%
BucketSeriesSkipChunks/1SeriesWith10000000Samples/100of10000000            195 ± 0%       172 ± 0%  -11.79%
BucketSeriesSkipChunks/1SeriesWith10000000Samples/10000000of10000000       543 ± 0%       445 ± 0%  -18.05%
```

2. Normal query

```
name                                                        old time/op    new time/op    delta
BucketSeries/1000000SeriesWith1Samples/1of1000000              451ms ± 0%     432ms ± 0%   -4.07%
BucketSeries/1000000SeriesWith1Samples/10of1000000             450ms ± 0%     450ms ± 0%   -0.09%
BucketSeries/1000000SeriesWith1Samples/1000000of1000000        3.82s ± 0%     4.27s ± 0%  +11.66%
BucketSeries/100000SeriesWith100Samples/1of10000000           36.7ms ± 0%    36.2ms ± 0%   -1.34%
BucketSeries/100000SeriesWith100Samples/100of10000000         36.1ms ± 0%    38.0ms ± 0%   +5.44%
BucketSeries/100000SeriesWith100Samples/10000000of10000000     321ms ± 0%     359ms ± 0%  +12.04%
BucketSeries/1SeriesWith10000000Samples/1of10000000           5.07ms ± 0%    0.17ms ± 0%  -96.56%
BucketSeries/1SeriesWith10000000Samples/100of10000000         4.80ms ± 0%    0.17ms ± 0%  -96.51%
BucketSeries/1SeriesWith10000000Samples/10000000of10000000    77.6ms ± 0%    81.6ms ± 0%   +5.08%

name                                                        old alloc/op   new alloc/op   delta
BucketSeries/1000000SeriesWith1Samples/1of1000000              117MB ± 0%      98MB ± 0%  -16.35%
BucketSeries/1000000SeriesWith1Samples/10of1000000             117MB ± 0%      98MB ± 0%  -16.35%
BucketSeries/1000000SeriesWith1Samples/1000000of1000000       1.31GB ± 0%    1.31GB ± 0%   +0.18%
BucketSeries/100000SeriesWith100Samples/1of10000000           10.4MB ± 0%     8.5MB ± 0%  -18.46%
BucketSeries/100000SeriesWith100Samples/100of10000000         10.4MB ± 0%     8.4MB ± 0%  -18.55%
BucketSeries/100000SeriesWith100Samples/10000000of10000000     126MB ± 0%     126MB ± 0%   +0.38%
BucketSeries/1SeriesWith10000000Samples/1of10000000           5.90MB ± 0%    0.22MB ± 0%  -96.19%
BucketSeries/1SeriesWith10000000Samples/100of10000000         5.90MB ± 0%    0.22MB ± 0%  -96.19%
BucketSeries/1SeriesWith10000000Samples/10000000of10000000    38.0MB ± 0%    38.0MB ± 0%   -0.00%

name                                                        old allocs/op  new allocs/op  delta
BucketSeries/1000000SeriesWith1Samples/1of1000000              1.26M ± 0%     0.76M ± 0%  -39.69%
BucketSeries/1000000SeriesWith1Samples/10of1000000             1.26M ± 0%     0.76M ± 0%  -39.69%
BucketSeries/1000000SeriesWith1Samples/1000000of1000000        12.1M ± 0%     12.1M ± 0%   -0.00%
BucketSeries/100000SeriesWith100Samples/1of10000000             126k ± 0%       76k ± 0%  -39.65%
BucketSeries/100000SeriesWith100Samples/100of10000000           126k ± 0%       76k ± 0%  -39.64%
BucketSeries/100000SeriesWith100Samples/10000000of10000000     1.21M ± 0%     1.21M ± 0%   -0.00%
BucketSeries/1SeriesWith10000000Samples/1of10000000              232 ± 0%       205 ± 0%  -11.64%
BucketSeries/1SeriesWith10000000Samples/100of10000000            231 ± 0%       205 ± 0%  -11.26%
BucketSeries/1SeriesWith10000000Samples/10000000of10000000      170k ± 0%      170k ± 0%   -0.00%

```