@yeya24 2020-10-28
After one more optimization, I think the performance is good now. That's the result:

1. SkipChunks

```
name                                                                  old time/op    new time/op    delta
BucketSeriesSkipChunks/1000000SeriesWith1Samples/1of1000000              423ms ± 0%     343ms ± 0%  -19.11%
BucketSeriesSkipChunks/1000000SeriesWith1Samples/10of1000000             496ms ± 0%     345ms ± 0%  -30.38%
BucketSeriesSkipChunks/1000000SeriesWith1Samples/1000000of1000000        3.39s ± 0%     2.85s ± 0%  -15.84%
BucketSeriesSkipChunks/100000SeriesWith100Samples/1of10000000           38.6ms ± 0%    26.6ms ± 0%  -31.15%
BucketSeriesSkipChunks/100000SeriesWith100Samples/100of10000000         36.1ms ± 0%    26.8ms ± 0%  -25.80%
BucketSeriesSkipChunks/100000SeriesWith100Samples/10000000of10000000     276ms ± 0%     257ms ± 0%   -6.70%
BucketSeriesSkipChunks/1SeriesWith10000000Samples/1of10000000           4.53ms ± 0%    0.12ms ± 0%  -97.27%
BucketSeriesSkipChunks/1SeriesWith10000000Samples/100of10000000         4.13ms ± 0%    0.12ms ± 0%  -96.98%
BucketSeriesSkipChunks/1SeriesWith10000000Samples/10000000of10000000    16.5ms ± 0%     0.4ms ± 0%  -97.48%

name                                                                  old alloc/op   new alloc/op   delta
BucketSeriesSkipChunks/1000000SeriesWith1Samples/1of1000000             98.2MB ± 0%    74.2MB ± 0%  -24.44%
BucketSeriesSkipChunks/1000000SeriesWith1Samples/10of1000000            98.2MB ± 0%    74.2MB ± 0%  -24.44%
BucketSeriesSkipChunks/1000000SeriesWith1Samples/1000000of1000000       1.00GB ± 0%    1.00GB ± 0%   +0.00%
BucketSeriesSkipChunks/100000SeriesWith100Samples/1of10000000           8.43MB ± 0%    6.03MB ± 0%  -28.48%
BucketSeriesSkipChunks/100000SeriesWith100Samples/100of10000000         8.43MB ± 0%    6.03MB ± 0%  -28.47%
BucketSeriesSkipChunks/100000SeriesWith100Samples/10000000of10000000    95.9MB ± 0%    95.9MB ± 0%   -0.00%
BucketSeriesSkipChunks/1SeriesWith10000000Samples/1of10000000           4.34MB ± 0%    0.17MB ± 0%  -96.00%
BucketSeriesSkipChunks/1SeriesWith10000000Samples/100of10000000         4.34MB ± 0%    0.17MB ± 0%  -96.00%
BucketSeriesSkipChunks/1SeriesWith10000000Samples/10000000of10000000    17.3MB ± 0%     0.7MB ± 0%  -96.06%

name                                                                  old allocs/op  new allocs/op  delta
BucketSeriesSkipChunks/1000000SeriesWith1Samples/1of1000000               760k ± 0%      510k ± 0%  -32.91%
BucketSeriesSkipChunks/1000000SeriesWith1Samples/10of1000000              760k ± 0%      510k ± 0%  -32.91%
BucketSeriesSkipChunks/1000000SeriesWith1Samples/1000000of1000000        8.04M ± 0%     8.04M ± 0%   +0.00%
BucketSeriesSkipChunks/100000SeriesWith100Samples/1of10000000            76.1k ± 0%     51.1k ± 0%  -32.86%
BucketSeriesSkipChunks/100000SeriesWith100Samples/100of10000000          76.1k ± 0%     51.1k ± 0%  -32.85%
BucketSeriesSkipChunks/100000SeriesWith100Samples/10000000of10000000      804k ± 0%      804k ± 0%   -0.00%
BucketSeriesSkipChunks/1SeriesWith10000000Samples/1of10000000              195 ± 0%       172 ± 0%  -11.79%
BucketSeriesSkipChunks/1SeriesWith10000000Samples/100of10000000            195 ± 0%       172 ± 0%  -11.79%
BucketSeriesSkipChunks/1SeriesWith10000000Samples/10000000of10000000       543 ± 0%       445 ± 0%  -18.05%
```

2. Normal series query

```
name                                                        old time/op    new time/op    delta
BucketSeries/1000000SeriesWith1Samples/1of1000000              451ms ± 0%     340ms ± 0%  -24.52%
BucketSeries/1000000SeriesWith1Samples/10of1000000             450ms ± 0%     345ms ± 0%  -23.46%
BucketSeries/1000000SeriesWith1Samples/1000000of1000000        3.82s ± 0%     3.85s ± 0%   +0.60%
BucketSeries/100000SeriesWith100Samples/1of10000000           36.7ms ± 0%    26.8ms ± 0%  -26.94%
BucketSeries/100000SeriesWith100Samples/100of10000000         36.1ms ± 0%    26.4ms ± 0%  -26.70%
BucketSeries/100000SeriesWith100Samples/10000000of10000000     321ms ± 0%     314ms ± 0%   -2.23%
BucketSeries/1SeriesWith10000000Samples/1of10000000           5.07ms ± 0%    0.15ms ± 0%  -97.03%
BucketSeries/1SeriesWith10000000Samples/100of10000000         4.80ms ± 0%    0.15ms ± 0%  -96.87%
BucketSeries/1SeriesWith10000000Samples/10000000of10000000    77.6ms ± 0%    70.5ms ± 0%   -9.23%

name                                                        old alloc/op   new alloc/op   delta
BucketSeries/1000000SeriesWith1Samples/1of1000000              117MB ± 0%      74MB ± 0%  -36.79%
BucketSeries/1000000SeriesWith1Samples/10of1000000             117MB ± 0%      74MB ± 0%  -36.80%
BucketSeries/1000000SeriesWith1Samples/1000000of1000000       1.31GB ± 0%    1.31GB ± 0%   +0.53%
BucketSeries/100000SeriesWith100Samples/1of10000000           10.4MB ± 0%     6.1MB ± 0%  -41.54%
BucketSeries/100000SeriesWith100Samples/100of10000000         10.4MB ± 0%     6.0MB ± 0%  -41.73%
BucketSeries/100000SeriesWith100Samples/10000000of10000000     126MB ± 0%     126MB ± 0%   +0.07%
BucketSeries/1SeriesWith10000000Samples/1of10000000           5.90MB ± 0%    0.22MB ± 0%  -96.19%
BucketSeries/1SeriesWith10000000Samples/100of10000000         5.90MB ± 0%    0.22MB ± 0%  -96.19%
BucketSeries/1SeriesWith10000000Samples/10000000of10000000    38.0MB ± 0%    38.0MB ± 0%   +0.00%

name                                                        old allocs/op  new allocs/op  delta
BucketSeries/1000000SeriesWith1Samples/1of1000000              1.26M ± 0%     0.51M ± 0%  -59.54%
BucketSeries/1000000SeriesWith1Samples/10of1000000             1.26M ± 0%     0.51M ± 0%  -59.53%
BucketSeries/1000000SeriesWith1Samples/1000000of1000000        12.1M ± 0%     12.1M ± 0%   +0.00%
BucketSeries/100000SeriesWith100Samples/1of10000000             126k ± 0%       51k ± 0%  -59.48%
BucketSeries/100000SeriesWith100Samples/100of10000000           126k ± 0%       51k ± 0%  -59.46%
BucketSeries/100000SeriesWith100Samples/10000000of10000000     1.21M ± 0%     1.21M ± 0%   +0.00%
BucketSeries/1SeriesWith10000000Samples/1of10000000              232 ± 0%       205 ± 0%  -11.64%
BucketSeries/1SeriesWith10000000Samples/100of10000000            231 ± 0%       205 ± 0%  -11.26%
BucketSeries/1SeriesWith10000000Samples/10000000of10000000      170k ± 0%      170k ± 0%   +0.00%

```